+++ src/utils.ts
@@ -1,5 +1,4 @@
-import { ChildProcess, spawn } from "child_process";
-import { TextDocument, Range, Position, OutputChannel, extensions, workspace, Event, Uri, FileSystemWatcher } from "vscode";
+import { TextDocument, Range, Position, extensions, workspace, Event, Uri, FileSystemWatcher } from "vscode";
 
 export const isEnclosedWithinCodeBlock: (
   document: TextDocument,
@@ -62,53 +61,6 @@ export const notEmpty = <T>(value: T | null | undefined): value is T => {
   return value !== null && value !== undefined;
 };
 
-export class CommandProcessExecution {
-  private readonly commandProcess: ChildProcess;
-  constructor(command: string, args?: string[], cwd?: string) {
-    this.commandProcess = spawn(
-      command,
-      args,
-      { cwd: cwd });
-  }
-
-  public async complete(): Promise<string> {
-    return new Promise<string>((resolve, reject) => {
-      const stdoutBuffer: Buffer[] = [];
-      const stderrBuffer: Buffer[] = [];
-      this.commandProcess.stdout!.on('data', chunk => stdoutBuffer.push(chunk));
-      this.commandProcess.stderr!.on('data', chunk => stderrBuffer.push(chunk));
-
-      this.commandProcess.once('close', () => {
-        const stdout = stdoutBuffer.toString();
-        const stderr = stderrBuffer.toString();
-        if (!stdout) {
-          reject(`Process returned an error:${stderr}`);
-        }
-        resolve(stdout);
-      });
-
-      this.commandProcess.once('error', error => {
-        reject(`Error occurred during process execution: ${error}`);
-      });
-    });
-
-  }
-
-  public async completeWithOutputChannel(outputChannel: OutputChannel): Promise<void> {
-    return new Promise((resolve, reject) => {
-      this.commandProcess.stdout!.on('data', chunk => { outputChannel.append(chunk.toString()); outputChannel.show(); });
-      this.commandProcess.stderr!.on('data', chunk => { outputChannel.append(chunk.toString()); outputChannel.show(); });
-      this.commandProcess.once('close', () => {
-        resolve();
-      });
-
-      this.commandProcess.once('error', error => {
-        reject(`Error occurred during process execution: ${error}`);
-      });
-    });
-  }
-}
-
 interface PythonExecutionDetails {
   pythonPath: string;
   onDidChangeExecutionDetails: Event<Uri | undefined>;
@@ -140,7 +92,7 @@ export const debounce = (fn: Function, wait: number) => {
   };
 };
 
-export const setupWatcherhHandler = (watcher: FileSystemWatcher, handler: Function): void => {
+export const setupWatcherHandler = (watcher: FileSystemWatcher, handler: Function): void => {
   watcher.onDidChange(() => handler());
   watcher.onDidCreate(() => handler());
   watcher.onDidDelete(() => handler());
